backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
ruff==0.1.6
spacy==3.7.2
rapidfuzz==3.5.2